Move class to /common
