+++ client/Player/ReactPlayerView.tsx
@@ -7,7 +7,7 @@ import {
 } from "../../common/EncounterState";
 import { PlayerViewSettings } from "../../common/PlayerViewSettings";
 import { PlayerViewState } from "../../common/PlayerViewState";
-import { StaticCombatantViewModel } from "../Combatant/StaticCombatantViewModel";
+import { StaticCombatantViewModel } from "../../common/StaticCombatantViewModel";
 import { getDefaultSettings } from "../Settings/Settings";
 import { PlayerView } from "./components/PlayerView";
